added remove all comments when deleting article. added light weight validation

+++ Bottles/Blog.Articles/Content/scripts/managearticles.js
@@ -3,15 +3,28 @@
 
   $('.delete-article').click(function () {
     var link = $(this),
-        id = link.data().id;
+        id = link.data().id,
+        proceedWithDeletion = confirm('Are you sure you want to delete (' + id + '). \r\nThis action cannot be undone.');
+
+    if (!proceedWithDeletion) return false;
+
     $.ajax({
       url: '/articles/manage',
       type: 'DELETE',
       data: { Id: id },
       dataType: 'json',
       success: function () {
-        link.closest('tr').fadeOut('slow');
-        messaging.raiseRemove($('.managearticlefilters:first').parent(), 'Article ('+ id +') has been successfully deleted.');
+        messaging.raiseRemove($('.managearticlefilters:first').parent(), 'Article (' + id + '), has been successfully deleted.');
+        $.ajax({
+          url: '/comments/delete-for-article',
+          type: 'DELETE',
+          data: { ArticleId: id },
+          dataType: 'json',
+          success: function () {
+            link.closest('tr').fadeOut('slow');
+            messaging.raiseRemove($('.managearticlefilters:first').parent(), 'Article (' + id + '), and article comments have been successfully deleted.');
+          }
+        });
       }
     });
   });
